requirements.txt
requests
beautifulsoup4
python-dotenv
Flask
Markdown
markdown2